Django_projects/basic_project/requirements.txt
bob-builder==0.0.5
boto==2.45.0
configparser==3.5.0
Django==1.9
djangorestframework==3.5.3
docopt==0.6.2
envoy==0.0.3
gunicorn==19.6.0
mysqlclient==1.3.9
requests==2.12.4
virtualenv==15.1.0
